Move omni and classic specific files into their respective main scripts
