+++ src/mbot_classic.h
@@ -23,4 +23,25 @@
 
 #include "config/mbot_classic_config.h"
 
+/**
+ * @brief Calculate the body velocity of a differential drive robot
+ *
+ * @param[in] wheel_left_vel  Velocity of the left wheel in rad/s
+ * @param[in] wheel_right_vel Velocity of the right wheel in rad/s
+ * @param[out] mbot_vel       Pointer to the structure where the calculated body velocity will be stored
+ * @return int                Returns 0 on success
+ */
+int mbot_calculate_diff_body_vel(float wheel_left_vel, float wheel_right_vel, serial_twist2D_t *mbot_vel);
+
+/**
+ * @brief Calculate the body velocity of a differential drive robot using an IMU for angular velocity
+ *
+ * @param[in] wheel_left_vel  Velocity of the left wheel in rad/s
+ * @param[in] wheel_right_vel Velocity of the right wheel in rad/s
+ * @param[in] imu             IMU data
+ * @param[out] mbot_vel       Pointer to the structure where the calculated body velocity will be stored
+ * @return int                Returns 0 on success
+ */
+int mbot_calculate_diff_body_vel_imu(float wheel_left_vel, float wheel_right_vel, serial_mbot_imu_t imu, serial_twist2D_t *mbot_vel);
+
 #endif /*MBOT_CLASSIC_H*/
